WIP: Testing HMMER refinement and clustering
It seems like, at this point, most of the difficult things are figured out. The problem is that either the distance measure (i.e., Google distance or Bray-curtis) or HDBSCAN is providing poor results. I will need to integrate another type of distance measure (i.e., IT-based rather than kmer based) to test the former, and if it's the latter that is the problem, I'll need to try to optimise parameters better.