Argument parsing (probably) finished
It's not pretty at times, but without having tested this on Linux yet I believe this works correctly. There might be an easier way to do this, but because the program has so many parameters and requires a bunch of programs, I think it's important that we handle things very explicitly and clearly so the program isn't quite as much of a nightmare trying to run for people unfamiliar with working with programs in command-line environments.