Most code modernisation complete
I probably need to figure out something a bit more elegant than hmm_grow, but for the time being it does still work. At this stage I believe that all the major code changes are in place for the base functionality of the program. Changes to be implemented are 1) modernise bechparse code, 2) modify hmm_grow, 3) remove hard-coded testing code, 4) update usage text to be more concise, 5) find a way to make clustering work better [This is the biggest and most important change... no real ideas of where to start with this]