msa_trim function for curation
Pretty cool function I think. Very simple but it has a lot of checks in it and it works how you'd expect it to. All that remains is to add a function to perform scoring. I will use blosum62 sum of pairs averaged across columns, and ideally through testing find some cutoff that works.