msa_score integrated
Scoring system is pretty average. Needs more testing to figure out if 0.20 is a good cutoff - all I can tell from the small testing I've done is that 0.15 is (probably) too low. Assuming this is a good cutoff and msa_score works as intended, then the program will be in a near-complete state (probably...). I will just need to tidy the code and add an extra filter at the end for cluster size - I think for program performance min_size needs to be 2 since changing this alters HDBSCAN's behaviour, but in reality I think a minimum cluster size of 3 makes more biological sense. This will be user specifiable of course.... Making the number of parameters a bit ridiculous - perhaps I'll consider some way of hiding many of these parameters?