MSA curation almost finalised
I've made a lot of small changes that are hard to individually list. Key changes are that I have a new function to remove outliers from MSAs using our dict structure (curate_msa_from_outlier_dict), and I am detecting outliers a bit better with some modifications to msa_outlier_detect. Some of these changes are a bit arbitrary in a sense, but I am basing it on manual inspection of real data and trying to figure out how to numerically codify the things which are obvious and apparent to my eye. I guess that means the next logical step is to use a neural network to do the prediction (I can dream...)