CANDID first "version"
Not sure how I'd number this. Maybe 0.9? It's not quite 1.0 since I'd like the benchparse code to be complete for that and there are still some minor changes require such as alphabet reduction for MSA scoring. Nonetheless, CANDID is operating in a way which is "good enough" for me to be happy with it. I've stripped out most code that was used for testing/is being tested but there are still a few flags throughout.