Major overhaul of hmm generating code
I've tried to make it resistant to most of my attempts to break the program. It's possible I've overlooked a few ways, but if you're going to get really creative to cause an unhandled error, then good on you. Otherwise the code is robust and does what it needs to do.